feature: h2 연결 설정
