+++ build.gradle
@@ -12,6 +12,7 @@ repositories {
 dependencies {
     implementation ('ch.qos.logback:logback-classic:1.4.12')
     implementation ('org.slf4j:slf4j-api:1.7.32')
+    implementation ('com.h2database:h2:2.3.230')
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation ('org.junit.jupiter:junit-jupiter')
 }
