+++ src/main/java/codesquad/Main.java
@@ -1,13 +1,16 @@
 package codesquad;
 
+import database.DatabaseConnector;
+import database.H2Server;
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.net.Socket;
+import java.sql.SQLException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import model.User;
 import org.slf4j.Logger;
-import repository.UserRepository;
+import repository.UserRepositoryMemory;
 import util.LoggerUtil;
 
 
@@ -20,20 +23,30 @@ public class Main {
 
         ExecutorService threadPool = Executors.newFixedThreadPool(THREAD_POOL_SIZE);
         // 유저 추가
-        UserRepository.getInstance().addUser( new User("a", "a", "a", "a"));
+        UserRepositoryMemory.getInstance().addUser( new User("a", "a", "a", "a"));
+        H2Server h2Server = new H2Server();
 
-        try (ServerSocket serverSocket = new ServerSocket(PORT)) {
-            logger.info("Server started on port " + PORT);
+        try {
+            h2Server.start();
+            // 데이터 베이스 커낵션 테스트
+            DatabaseConnector.getInstance().getConnection();
+            try (ServerSocket serverSocket = new ServerSocket(PORT)) {
+                logger.info("Server started on port " + PORT);
 
-            while (true) {
-                Socket clientSocket = serverSocket.accept();
-                logger.info("Client  connected");
+                while (true) {
+                    Socket clientSocket = serverSocket.accept();
+                    logger.info("Client  connected");
 
-                // 클라이언트 요청을 처리하기 위해 스레드 풀에 작업 제출
-                threadPool.submit(new ClientThread(clientSocket));
+                    // 클라이언트 요청을 처리하기 위해 스레드 풀에 작업 제출
+                    threadPool.submit(new ClientThread(clientSocket));
+                }
+            } catch (IOException e) {
+                logger.error("Could not start server: {}", e.getMessage());
             }
-        } catch (IOException e) {
-            logger.error("Could not start server: {}", e.getMessage());
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }finally {
+            h2Server.stop();
         }
     }
